Removes mappers package
